\name{ExtractTimePeriod}
\alias{ExtractTimePeriod}

\title{
Extract a subset of a mobility matrix constrained to a specified time period.
}
\description{
Extract a subset of a mobility matrix constrained to a specified time period.
}
\usage{
ExtractTimePeriod(tstart, tend, out)
}

\arguments{
  \item{tstart}{
A string specifying the starting time of the interval to be extracted. For example: "2016-01-29 00:00:00 EDT".
  }
  \item{tend}{
A string specifying the ending time of the interval to be extracted. For example: "2016-01-29 23:59:00 EDT".
}
  \item{out}{
Mobility matrix as output by MobilityFeatures().
  }
}
\details{
}
\value{
A mobility matrix.
}
\references{
}
\author{
Ian Barnett
}
\note{
}


\seealso{
}
\examples{
}
\keyword{ manip }
