\name{plotlimits}
\alias{plotlimits}

\title{
Finds the x and y limits of a mobility matrix.
}
\description{
Finds the x and y limits of a mobility matrix. This is useful for plotting purposes in the plot.flights() function.
}
\usage{
plotlimits(mat)
}

\arguments{
  \item{mat}{
  Mobility matrix.
}
}
\details{
The mobility matrix (mat) is output from the MobilityFeatures() function.
}
\value{
 A list with two values, xrang and yrang. Both are vectors of length 2, a minimum and maximum value specifying a range.
}
\references{
}
\author{
Ian Barnett
}
\note{
}


\seealso{
\code{\link{plot.flights}}
}
\examples{
}
\keyword{ attribute}
